+++ owncloudApp/src/test/java/com/owncloud/android/presentation/viewmodels/settings/SettingsSecurityViewModelTest.kt
@@ -20,8 +20,8 @@
 
 package com.owncloud.android.presentation.viewmodels.settings
 
-import com.owncloud.android.R
 import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
+import com.owncloud.android.enums.LockEnforcedType
 import com.owncloud.android.presentation.ui.security.PassCodeActivity
 import com.owncloud.android.presentation.ui.security.PatternActivity
 import com.owncloud.android.presentation.ui.settings.fragments.SettingsSecurityFragment
@@ -139,28 +139,18 @@ class SettingsSecurityViewModelTest : ViewModelTest() {
 
     @Test
     fun `set pref is security enforced enabled - ok - true`() {
-        every { contextProvider.getBoolean(any()) } returns true
+        every { contextProvider.getInt(any()) } returns LockEnforcedType.EITHER_ENFORCED.ordinal
 
-        securityViewModel.isSecurityEnforcedEnabled().apply {
-            assertTrue(this)
-        }
+        val result = securityViewModel.isSecurityEnforcedEnabled()
+        assertTrue(result)
 
-        verify(exactly = 1) {
-            contextProvider.getBoolean(R.bool.lock_enforced)
-        }
     }
 
     @Test
     fun `set pref is security enforced enabled - ok - false`() {
-        every { contextProvider.getBoolean(any()) } returns false
+        every { contextProvider.getInt(any()) } returns LockEnforcedType.DISABLED.ordinal
 
-        securityViewModel.isSecurityEnforcedEnabled().apply {
-            assertFalse(this)
-        }
-
-        verify(exactly = 1) {
-            contextProvider.getBoolean(R.bool.lock_enforced)
-        }
+        val result = securityViewModel.isSecurityEnforcedEnabled()
+        assertFalse(result)
     }
-
 }